Add base class for AWS Lambda upgrades
Add a new base class for processing upgrades for AWS Lambda.
