+++ src/DotNetBumper.Core/WellKnownFileNames.cs
@@ -11,6 +11,7 @@ internal static class WellKnownFileNames
     public const string PublishProfiles = "*.pubxml";
     public const string SolutionFiles = "*.sln";
 
+    public const string AwsLambdaToolsDefaults = "aws-lambda-tools-defaults.json";
     public const string DirectoryBuildProps = "Directory.Build.props";
     public const string GlobalJson = "global.json";
     public const string ToolsManifest = "dotnet-tools.json";
